Move data and models to apf
