+++ flyllm/plotting.py
@@ -10,11 +10,11 @@ from flyllm.config import (
     featglobal, featrelative, kpvision_other,
     nglobal, nrelative, nkptouch, nfeatures
 )
-from flyllm.data import get_batch_idx, load_and_filter_data, split_data_by_id, select_bin_edges, get_real_flies
-from flyllm.features import compute_features, get_sensory_feature_idx, zscore, unzscore
+from apf.data import get_batch_idx, load_and_filter_data, split_data_by_id, select_bin_edges, get_real_agents
+from flyllm.features import compute_features, zscore, unzscore
 from flyllm.pose import FlyExample
-from flyllm.models import criterion_wrapper
-from flyllm.utils import npindex
+from apf.models import criterion_wrapper
+from apf.utils import npindex
 from flyllm.io import read_config
 
 
@@ -122,7 +122,7 @@ def plot_fly(pose=None, kptidx=keypointidx, skelidx=skeleton_edges, fig=None, ax
     assert (skelidx is not None)
 
     fig, ax, isnewaxis = set_fig_ax(fig=fig, ax=ax)
-    isreal = get_real_flies(pose[:, :, np.newaxis])
+    isreal = get_real_agents(pose[:, :, np.newaxis])
 
     hkpts = None
     hedges = None
